chore: Move /user/welcome to /redesign
