+++ src/app/(proper_react)/redesign/(authenticated)/user/welcome/EnterInfo.module.scss
@@ -1,4 +1,4 @@
-@import "../../../../tokens";
+@import "../../../../../tokens";
 
 .stepContent {
   display: flex;
