+++ src/app/(proper_react)/redesign/(authenticated)/user/welcome/FindExposures.module.scss
@@ -1,4 +1,4 @@
-@import "../../../../tokens";
+@import "../../../../../tokens";
 
 .progressLabel {
   margin: $spacing-xl 0 $spacing-lg;
