+++ src/app/(proper_react)/redesign/(authenticated)/user/welcome/Onboarding.test.tsx
