+++ src/app/(proper_react)/redesign/(authenticated)/user/welcome/View.module.scss
@@ -1,4 +1,4 @@
-@import "../../../../tokens";
+@import "../../../../../tokens";
 
 .wrapper {
   --backgroundImageWidth: 400px;
